added JSONWrapper class (broken for now)
